+++ HISTORY.md
@@ -23,6 +23,7 @@
     * new "this script" primitive, experimental
     * new localization extension primitives in the "ide" category, hyperized
     * new extension primitive for importing a costume from a url
+    * new extension primitive for querying all variable names accessible from a specified scope (global, sprite, script)
     * new support for setting the translation via the API
     * new "Tad" and "Jahrd" costume series, thanks, Meghan and Brian!
 * **Notable Changes:**
@@ -59,6 +60,7 @@
 * threads: added support for "defaults" selector on primitives
 * threads: added support for "editables" selector on primitives
 * threads, byob: added support for "menus" selector on primitives
+* extensions: new extension primitive for querying all variable names accessible from a specified scope (global, sprite, script)
 
 ### 2022-06-28
 * blocks, byob, threads: new "menus" selector for block attributes